Make --language optional
Take the CLI, then the Pragma, then fall back to the most common
code-block style in the file. If none of that works, throw an exception.

The exception has a really nice `displayException`, but I don't think
it'll be used when the program is crashed. We'll see if it's worth
working around Haskell's odd choice here.